+++ CMakeLists.txt
@@ -4,10 +4,11 @@ project(LeDernierMorkid)
 set(EXECUTABLE_OUTPUT_PATH ${PROJECT_BINARY_DIR}/bin)
 
 file(GLOB SRC "src/*.cpp")
+file(GLOB GRAPHICSSRC "src/Graphics/*.cpp")
 file(GLOB TERRAINSRC "src/Terrain/*.cpp" "src/Terrain/Hlms/*.cpp" "src/Terrain/Hlms/PbsListener/*.cpp")
 file(GLOB COLLISIONSRC "Collision/src/*.cpp")
 file(GLOB COMMONSRC "Common/src/*.cpp" "Common/src/Threading/*.cpp" "Common/src/System/*.cpp")
-add_executable(LeDernierMorkid ${SRC} ${TERRAINSRC} ${COLLISIONSRC} ${COMMONSRC})
+add_executable(LeDernierMorkid ${SRC} ${GRAPHICSSRC} ${TERRAINSRC} ${COLLISIONSRC} ${COMMONSRC})
 
 add_custom_target(run COMMAND LeDernierMorkid DEPENDS LeDernierMorkid WORKING_DIRECTORY ${PROJECT_BINARY_DIR}/bin)
